Video Scrubber, Resources, Icons
We built a fully functional video progress bar that works on the pause menu. It shows where you are in the video and how much is left. The breathing tool works in the pause menu as well. We finished the resource page so it plays all the videos that are needed, just need one more video from Julian. We updated icons and UI for a lot of the pages so it's easier to follow.